requirements.txt
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
filelock==3.14.0
identify==2.5.36
idna==3.7
nodeenv==1.8.0
platformdirs==4.2.2
pre-commit==3.7.1
pyTelegramBotAPI==4.17.0
PyYAML==6.0.1
requests==2.31.0
telebot==0.0.5
urllib3==2.2.1
virtualenv==20.26.2
